Fixed some issues with tests found by CircleCI
