+++ src/test/scala/com/phasmidsoftware/table/AnalysisSpec.scala
@@ -39,39 +39,6 @@ class AnalysisSpec extends AnyFlatSpec with Matchers {
     }
   }
 
-  it should "analyze the complete crime file" in {
-    val crimeFile = "../examples/crime/2023-01-metropolitan-street.csv"
-
-    implicit object validityRawRow extends Validity[RawRow] {
-      def isValid(r: RawRow): Boolean = ! {
-        val latitude: Double = r("latitude").get.toDoubleOption.getOrElse(55)
-        val longitude: Double = r("longitude").get.toDoubleOption.getOrElse(1)
-        val lsoaCode = r("LSOA code").getOrElse("")
-        CrimeLocation.isValid(longitude, latitude, lsoaCode)
-      }
-    }
-
-    // Set up the source
-    val sy: IO[Source] = IO.fromTry(for (u <- FP.resource[Analysis](crimeFile)) yield Source.fromURL(u))
-
-    val fraction = 1
-    // Set up the parser (we set the predicate only for demonstration purposes)
-    val parser: RawTableParser = RawTableParser().setPredicate(TableParser.sampler(fraction))
-
-    EvaluateIO.check(parser.parse(sy), Timeout(Span(10, Seconds))) {
-      case t@HeadedTable(r, _) =>
-        val q = t.filterValid
-        Analysis(q) match {
-          case a@Analysis(_, 12, _) =>
-            println(s"Crime analysis: $a")
-            r take 10 foreach println
-          case _ =>
-            println(s"Not good analysis")
-            fail("didnt match")
-        }
-    }
-  }
-
   behavior of "Column"
 
   it should "makeNumeric" in {
